Add study name to study URL
