+++ src/components/network/network-explorer.js
@@ -50,7 +50,9 @@ const NetworkExplorer = (props) => {
     };
 
     useEffect(() => {
-        setFilteredVoltageLevels(props.network.getVoltageLevels().sort(voltageLevelComparator))
+        if (props.network) {
+            setFilteredVoltageLevels(props.network.getVoltageLevels().sort(voltageLevelComparator))
+        }
     }, [props.network]);
 
     useEffect(() => {
@@ -108,11 +110,11 @@ const NetworkExplorer = (props) => {
 };
 
 NetworkExplorer.defaultProps = {
-    network: new Network()
+    network: null
 };
 
 NetworkExplorer.propTypes = {
-    network: PropTypes.instanceOf(Network).isRequired,
+    network: PropTypes.instanceOf(Network),
     onVoltageLevelClick: PropTypes.func
 };
