+++ src/components/study-pane.js
@@ -5,9 +5,11 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-import React from "react";
+import React, {useEffect} from "react";
 
 import {useDispatch, useSelector} from "react-redux";
+
+import {useRouteMatch} from 'react-router-dom';
 
 import Grid from "@material-ui/core/Grid";
 import {makeStyles} from "@material-ui/core/styles";
@@ -15,8 +17,23 @@ import {makeStyles} from "@material-ui/core/styles";
 import NetworkExplorer from "./network/network-explorer";
 import NetworkMap from "./network/network-map";
 import SingleLineDiagram from "./single-line-diagram";
-import {fetchVoltageLevelDiagram} from "../utils/rest-api";
-import {loadVoltageLevelDiagramSuccess, removeVoltageLevelDiagram} from "../redux/actions";
+import {
+    fetchLinePositions,
+    fetchLines,
+    fetchSubstationPositions,
+    fetchSubstations,
+    fetchVoltageLevelDiagram
+} from "../utils/rest-api";
+import {
+    closeStudy,
+    loadGeoDataSuccess,
+    loadNetworkSuccess,
+    loadVoltageLevelDiagramSuccess,
+    openStudy,
+    removeVoltageLevelDiagram
+} from "../redux/actions";
+import Network from "./network/network";
+import GeoData from "./network/geo-data";
 
 const useStyles = makeStyles(theme => ({
     main: {
@@ -42,6 +59,51 @@ const StudyPane = () => {
 
     const classes = useStyles();
 
+    const match = useRouteMatch("/studies/:studyName");
+    const studyName = match.params.studyName;
+
+    useEffect(() => {
+        dispatch(openStudy(studyName));
+        loadNetwork(studyName);
+        loadGeoData(studyName);
+
+        return function () {
+            dispatch(closeStudy());
+        }
+    }, [studyName]);
+
+    function loadNetwork(studyName) {
+        console.info(`Loading network of study '${studyName}'...`);
+
+        const substations = fetchSubstations(studyName);
+
+        const lines = fetchLines(studyName);
+
+        Promise.all([substations, lines])
+            .then(values => {
+                const network = new Network();
+                network.setSubstations(values[0]);
+                network.setLines(values[1]);
+                dispatch(loadNetworkSuccess(network));
+            });
+    }
+
+    function loadGeoData(studyName) {
+        console.info(`Loading geo data of study '${studyName}'...`);
+
+        const substationPositions = fetchSubstationPositions(studyName);
+
+        const linePositions = fetchLinePositions(studyName);
+
+        Promise.all([substationPositions, linePositions])
+            .then(values => {
+                const geoData = new GeoData();
+                geoData.setSubstationPositions(values[0]);
+                geoData.setLinePositions(values[1]);
+                dispatch(loadGeoDataSuccess(geoData));
+            });
+    }
+
     function showVoltageLevelDiagram(voltageLevelId, voltageLevelName) {
         // load svg
         fetchVoltageLevelDiagram(study.name, voltageLevelId, useName)
@@ -55,6 +117,7 @@ const StudyPane = () => {
     }
 
     return (
+        study &&
         <Grid container className={classes.main}>
             <Grid item xs={12} md={2} key="explorer">
                 <NetworkExplorer network={study.network}
